updated vending machine project
i've added a bit more to the GUI for the project. It's a rough version, so feel free to edit/clean the code up if you'd like. I've been working on this for a couple of hours. I'm gonna call it quits for the night. I couldn't figure out how to use an array and then have it appear has JButtons on the left hand side. I've marked where it is within the code. let me know if you have any questions. thanks